Move `color()` function parsing for XYZ
