+++ src/parser.rs
@@ -12,7 +12,7 @@ use nom::{
 
 use crate::{
     cmyk::parse_cmyk_color, hsl::parse_hsl_color, hsv::parse_hsv_color, hwb::parse_hwb_color,
-    lab::parse_lab_color, lch::parse_lch_color, rgb::parse_rgb_color,
+    lab::parse_lab_color, lch::parse_lch_color, rgb::parse_rgb_color, xyz::parse_xyz_color,
 };
 use crate::{named::NAMED_COLORS, Color};
 
@@ -146,37 +146,14 @@ where
     }
 }
 
-fn parse_css_color_fn(input: &str) -> IResult<&str, Color> {
-    all_consuming(parse_cie_xyz65_color_space)(input.trim())
-}
-
-// CSS Color 4 defines separate D65-adapted (`xyz-d65`, or just `xyz`) and D5-adapted (`xyz-d50`)
-// color spaces.  Currently, `pastel` does not support chromatic adaptation, and only uses the D65
-// illuminant, so we only support the `xyz-d65` color space here.
-fn parse_cie_xyz65_color_space(input: &str) -> IResult<&str, Color> {
-    fn xyz_components(input: &str) -> IResult<&str, Color> {
-        let (input, x) = number_or_percentage(input, 1.0)?;
-        let (input, _) = space1(input)?;
-        let (input, y) = number_or_percentage(input, 1.0)?;
-        let (input, _) = space1(input)?;
-        let (input, z) = number_or_percentage(input, 1.0)?;
-
-        let c = Color::from_xyz(x, y, z, 1.0);
-        Ok((input, c))
-    }
-
-    let xyz_name = alt((tag_no_case("xyz-d65"), tag_no_case("xyz")));
-    css_color_function(xyz_name, xyz_components)(input)
-}
-
 pub fn parse_color(input: &str) -> Option<Color> {
     alt((
         parse_rgb_color,
         parse_hsl_color,
-        parse_css_color_fn,
         parse_hsv_color,
         parse_hwb_color,
         all_consuming(parse_gray),
+        parse_xyz_color,
         parse_lab_color,
         parse_lch_color,
         parse_cmyk_color,
@@ -370,7 +347,7 @@ fn parse_color_srgb_string() {
 }
 
 #[test]
-fn parse_xyz65_color_space_syntax() {
+fn parse_color_xyz_d65_string() {
     fn xyz(x: f64, y: f64, z: f64) -> Color {
         Color::from_xyz(x, y, z, 1.0)
     }
@@ -379,76 +356,14 @@ fn parse_xyz65_color_space_syntax() {
         Some(xyz(0.3, 0.5, 0.7)),
         parse_color("color(xyz-d65 0.3 0.5 0.7)")
     );
-
-    assert_eq!(
-        Some(xyz(0.950_470, 1.0, 1.088_830)),
-        parse_color("color(xyz-d65 0.950470 1 1.088830)")
-    );
-
-    assert_eq!(
-        Some(xyz(-0.004, 1.007, -1.2222)),
-        parse_color("color(xyz-d65 -0.004 1.007000 -1.2222)")
-    );
-
-    // percentages are allowed
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(xyz-d65 30% 50% 70%)")
-    );
-    // numbers and percentages can be mixed
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(xyz-d65 0.3 50% 0.7)")
-    );
-
-    // `xyz` is equivalent to `xyz-d65`
     assert_eq!(
         Some(xyz(0.3, 0.5, 0.7)),
         parse_color("color(xyz 0.3 0.5 0.7)")
     );
-
-    // extra spaces are allowed
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(xyz-d65  0.3   0.5    0.7)")
-    );
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(   xyz-d65   0.3 0.5 0.7)")
-    );
-
-    // color space name is case-insensitive
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(XYZ 0.3 0.5 0.7)")
-    );
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(Xyz 0.3 0.5 0.7)")
-    );
-    assert_eq!(
-        Some(xyz(0.3, 0.5, 0.7)),
-        parse_color("color(xyz-D65 0.3 0.5 0.7)")
-    );
-
-    // alpha is supported
     assert_eq!(
         Some(Color::from_xyz(0.3, 0.5, 0.7, 0.9)),
         parse_color("color(xyz-d65 0.3 0.5 0.7 / 0.9)")
     );
-
-    // not enough parameters
-    assert_eq!(None, parse_color("color(xyz-d65 0.3 0.5)"));
-    // too many parameters
-    assert_eq!(None, parse_color("color(xyz-d65 0.3 0.5 0.7 1.0)"));
-    // comma separators not allowed
-    assert_eq!(None, parse_color("color(xyz-d65 0.3, 0.5, 0.7)"));
-}
-
-#[test]
-fn parse_xyz50_color_space_syntax() {
-    // The D50-adapted `xyz-d50` color space is not currently supported.
-    assert_eq!(None, parse_color("color(xyz-d50 0.3 0.5 0.7)"));
 }
 
 #[test]
